[radiation] update & moved to aboveground
