+++ ResourceLib/AboveGround/SolarRadiation/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+.. include:: ./SolarRadiation.md
+"""
+
+from .SolarRadiation import SolarRadiation
